fix(indexer): widen the funded watermark column to bigint
The boot reconciliation stamps reconciled rows with a block-end sentinel
logIndex (Number.MAX_SAFE_INTEGER) that a 32-bit INTEGER column rejects,
rolling back the whole atomic persist and re-failing on every restart:
the reviewer-identified wedge on exactly the pre-feature upgrade path.
First half of the REQUEST-CHANGES round; the branch pauses here (the
receive product line continues in the pripay repo), so the remaining
findings (reconciliation set-not-add, tail unit coverage) are recorded in
the review, not fixed.

+++ apps/indexer/src/schema.sql
@@ -179,7 +179,10 @@ ALTER TABLE portal_announcements ADD COLUMN IF NOT EXISTS funded_amount TEXT;   
 ALTER TABLE portal_announcements ADD COLUMN IF NOT EXISTS funded_tx_hash TEXT;        -- first qualifying transfer
 ALTER TABLE portal_announcements ADD COLUMN IF NOT EXISTS funded_at BIGINT;           -- unix seconds of that transfer
 ALTER TABLE portal_announcements ADD COLUMN IF NOT EXISTS funded_block BIGINT;        -- watermark half 1
-ALTER TABLE portal_announcements ADD COLUMN IF NOT EXISTS funded_log_index INTEGER;   -- watermark half 2
+-- BIGINT, not INTEGER: the boot reconciliation's watermark uses a
+-- block-end sentinel (Number.MAX_SAFE_INTEGER) that a 32-bit column would
+-- reject, rolling back every persist and wedging ingest permanently.
+ALTER TABLE portal_announcements ADD COLUMN IF NOT EXISTS funded_log_index BIGINT;    -- watermark half 2
 CREATE INDEX IF NOT EXISTS portal_destination_idx ON portal_announcements (destination);
 
 -- The funded tail's own scan cursor (ingest_cursor's twin): lags the pool
